modularized borderColor and buttonColor functions into helper.js and added background color hover styling to dev cards buttons
