+++ src/styles/buttons.scss
@@ -54,6 +54,78 @@ $platinum: #ecebea;
   // box-shadow: 2px 2px 1px grey;
 }
 
+//Resume button
+.dev-card__resume {
+  display: flex;
+  justify-content: space-around;
+  width: 250px;
+  padding: 5px 15px;
+  margin: 5px;
+  border: transparent;
+  border-radius: 8px;
+  background-color: $gray;
+  color: white;
+  cursor: pointer;
+  font-family: "Poppins", sans-serif;
+  font-weight: 600;
+  font-size: 18px;
+  letter-spacing: 3px;
+  // background-image: linear-gradient(#444c4c, #747e80);
+  
+  
+}
+//Resume button hover
+.dev-card__resume:hover {
+  background-color: $orange;
+  color: #120000; 
+}
+.fsButtonYellow {
+  display: flex;
+  justify-content: space-around;
+  width: 250px;
+  padding: 5px 15px;
+  margin: 5px;
+  border: transparent;
+  border-radius: 8px;
+  background-color: $gray;
+  color: white;
+  cursor: pointer;
+  font-family: "Poppins", sans-serif;
+  font-weight: 600;
+  font-size: 18px;
+  letter-spacing: 3px;
+}
+
+.fsButtonYellow:hover {
+  background-color: $mustard;
+  color: #120000; 
+}
+//this className cooresponds to one of the return values of buttonColor 
+.feButtonBlue {
+  display: flex;
+  justify-content: space-around;
+  width: 250px;
+  padding: 5px 15px;
+  margin: 5px;
+  border: transparent;
+  border-radius: 8px;
+  background-color: $gray;
+  color: white;
+  cursor: pointer;
+  font-family: "Poppins", sans-serif;
+  font-weight: 600;
+  font-size: 18px;
+  letter-spacing: 3px;
+}
+//this className cooresponds to one of the return values of buttonColor 
+.feButtonBlue:hover {
+  background-color: $blue;
+  color: #120000; 
+}
+
+//Personal Site button
+
+
 //Podcast Player section
 .pod-player {
   display: flex;
